Cleanup More Screen. Add additional tabs. Cleanup some onboarding screens now used by the more screen
