+++ PittsburghCityBridges/Constants/TextCopy.swift
@@ -22,12 +22,13 @@ No representation is made or warranty given as to data content, route practicabi
     static let onBoardingIntroFirstPart = """
 This Pittsburgh City Bridges app contains information on over 140 bridges publicly available from the City of Pittsburgh Bridges Open Data database.
 
-This app does not have the big bridges that cross any of the three rivers because they are not in the City of Pittsburgh Bridges Open Data database.
+This app was created by one iOS developer as a so-called portfolio app to showcase using the latest iOS technology. The City of Pittsburgh Bridges database was selected because there "wasn't an App for that". App development work began in October 2021, never knowing the Fern Hollow Bridge would collapse some months later in January 2022.
 
-Instead the bridges in this app cross roads, ravines, and streams and are understood to be city maintained bridges.
+The bridges in this app cross roads, ravines, and streams and are understood to be city maintained bridges.
 
-Search bridges by neighborhood, name and year built.
+This app does not have the major bridges that cross any of the three rivers because they are not in the City of Pittsburgh Bridges Open Data database.
 
+This app is provided to the community free of charge.
 """
     
     static let onboardingOpenDataSource = """
@@ -37,9 +38,9 @@ This Pittsburgh City Bridges App is a data viewer for bridge information contain
     static let onBoardingCloseScreen = """
 App Maker makes no guarantee about accuracy of data contained in the City of Pittsburgh Bridges publicly available dataset.
 
-App Maker cannot verify the GPS coordinates or even existence of all 140 bridges. Use at your own risk.
+App Maker cannot verify the GPS coordinates or even existence of all 140 bridges.
 
-App Maker has found this to be an educational and even entertaining app. It is presented to the community free of charge.
+Use at your own risk.
 """
     
     static let onBoardingCloseScreenButton = """
@@ -47,11 +48,11 @@ Close
 """
     
     static let onBoardingCollapsedBridge = """
-    This is the Pittsburgh bridge that collapsed on January 28, 2022. It crosses the Fern Hollow creek and is commonly called the Fern Hollow Bridge. To find it in this app search for Forbes Avenue Frick Park Bridge in the List or Photos views.
+    This is the Pittsburgh bridge that collapsed on January 28, 2022. It crosses the Fern Hollow creek and is commonly called the Fern Hollow Bridge. It is named the Forbes Avenue Frick Park Bridge in the City of Pittsburgh Bridges database.
     """
     
     static let wprdcDescription = """
-    "... The Western Pennsylvania Regional Data Center supports key community initiatives by making public information easier to find and use. The Data Center provides a technological and legal infrastructure for data sharing to support a growing ecosystem of data providers and data users. The Data Center maintains Allegheny County and the City of Pittsburgh’s open data portal, and provides a number of services to data publishers and users. The Data Center also hosts datasets from these and other public sector agencies, academic institutions, and non-profit organizations. The Data Center is managed by the University of Pittsburgh’s Center for Social and Urban Research, and is a partnership of the University, Allegheny County and the City of Pittsburgh ..."
+    From the website "...The Western Pennsylvania Regional Data Center supports key community initiatives by making public information easier to find and use. The Data Center provides a technological and legal infrastructure for data sharing to support a growing ecosystem of data providers and data users. The Data Center maintains Allegheny County and the City of Pittsburgh’s open data portal, and provides a number of services to data publishers and users. The Data Center also hosts datasets from these and other public sector agencies, academic institutions, and non-profit organizations. The Data Center is managed by the University of Pittsburgh’s Center for Social and Urban Research, and is a partnership of the University, Allegheny County and the City of Pittsburgh ..."
     """
