--- PittsburghCityBridges/Views/Onboarding/OnBoardingContentView.swift
@@ -1,31 +1,0 @@
-//
-//  OnBoardingScreenView.swift
-//  PittsburghCityBridges
-//
-//  Created by MAKinney on 1/16/22.
-//
-
-import SwiftUI
-
-struct OnBoardingContentView: View {
- 
-    var body: some View {
-        TabView {
-            OnBoardingIntroScreen()
-//            OnBoardingBridgePhotosScreen()
-//            OnBoardingBrowseScreen()
-//            OnBoardingMapScreen()
-//            OnBoardingSortAndSearchScreen()
-//            OnboardingCollapsedBridgeScreen()
-//            OnBoardingCloseScreen()
-        }
-        .background(Color.pbBgnd)
-        .tabViewStyle(.page(indexDisplayMode: .always))
-    }
-}
-
-struct OnBoardingScreenView_Previews: PreviewProvider {
-    static var previews: some View {
-        OnBoardingContentView()
-    }
-}
