+++ PittsburghCityBridges/Views/Onboarding/OnBoardingIntroScreen.swift
@@ -9,16 +9,21 @@ import SwiftUI
 
 struct OnBoardingIntroScreen: View {
     var body: some View {
-        VStack(alignment: .leading) {
-            Spacer()
-            Text(AppTextCopy.onBoardingIntroFirstPart)
-            .padding()
-            Spacer()
+        ZStack {
+            Color.pbBgnd
+            ScrollView {
+                VStack(alignment: .leading) {
+                    Text(AppTextCopy.onBoardingIntroFirstPart)
+                        .padding()
+                    Spacer()
+                }
+                .font(UIDevice.current.userInterfaceIdiom == .phone ? .subheadline : .title2)
+                .frame(maxWidth: UIDevice.current.userInterfaceIdiom == .phone ? 350 : 500)
+                .multilineTextAlignment(.leading)
+                .background(Color.pbBgnd)
+                .foregroundColor(.pbTextFnd)
+            }
         }
-        .font(UIDevice.current.userInterfaceIdiom == .phone ? .subheadline : .title2)
-        .frame(maxWidth: UIDevice.current.userInterfaceIdiom == .phone ? 350 : 500)
-        .multilineTextAlignment(.leading)
-        .foregroundColor(.pbTextFnd)
     }
 }
