+++ PittsburghCityBridges/Views/Onboarding/OnboardingCollapsedBridgeScreen.swift
@@ -9,25 +9,26 @@ import SwiftUI
 
 struct OnboardingCollapsedBridgeScreen: View {
     var body: some View {
-        VStack {
-            Spacer()
-            Text(AppTextCopy.onBoardingCollapsedBridge)
-                .font(UIDevice.current.userInterfaceIdiom == .phone ? .subheadline : .title2)
-                .padding(.vertical)
-            Image("frickParkBridge")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .clipped()
-                .cornerRadius(PCBImage.cornerRadius)
-                .background (
-                    RoundedRectangle(cornerRadius: PCBImage.cornerRadius)
-                        .stroke(Color.pbTextFnd, lineWidth: 4)
-                )
-            Spacer()
-        }
-        .frame(maxWidth: UIDevice.current.userInterfaceIdiom == .phone ? 350 : 700)
-        .background(Color.pbBgnd)
+        ZStack {
+            Color.pbBgnd
+            VStack {
+                Text(AppTextCopy.onBoardingCollapsedBridge)
+                    .font(UIDevice.current.userInterfaceIdiom == .phone ? .subheadline : .title2)
+                    .padding(.vertical)
+                Image("frickParkBridge")
+                    .resizable()
+                    .aspectRatio(contentMode: .fit)
+                    .clipped()
+                    .cornerRadius(PCBImage.cornerRadius)
+                    .background (
+                        RoundedRectangle(cornerRadius: PCBImage.cornerRadius)
+                            .stroke(Color.pbTextFnd, lineWidth: 4)
+                    )
+                Spacer()
+            }
+            .frame(maxWidth: UIDevice.current.userInterfaceIdiom == .phone ? 350 : 700)
         .foregroundColor(.pbTextFnd)
+        }
     }
 }
